Location of ID within path can be specified by %s
Sometimes (but only with subscription) the ID for the object is in the
middle of the path.  There needed to be a way to specify the proper
location.

The solution was to append the ID by default, but when a %s was present
in the path, use sprintf() to insert the id into the path appropriately.

For an example of what I'm talking about, see:
 * https://stripe.com/docs/api
For convenience, it's:
/v1/customers/{CUSTOMER_ID}/subscription