Deletes now accept parameters
Deletion of a subscription object can include parameters.  However,
there was not support for supplying arguments to a delete at all.

This was a weird one to resolve, as the Cake API's aren't set up to
handle something like this.

I ended up just adding an additional parameter to the delete methods and
bundling the arguments as a property of the model.  Not ideal, but it
works.