StripeSource::read supports list queries
StripeSource::read was previously written to retrieve single records.
This did not permit it to do full scale find operations, such as with
invoices.  This was particularly problematic when you wanted to pass
query params as well.

We now detect whether this is a find for a single record or multiple
records and perform the query accordingly.

Additionally, the read method set the id on the model.  This is actually
inconsistent with CakePHP's ordinary models; read is the only operation
that does this, and it handles that task itself.  We have removed this
functionality.